pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "venv"
version = "0.1.0"
description = "Cultural Question Answering using Llama 3-8B"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    {name = "Omar El Ansary"},
    {name = "Hussam Chihadeh"}
]
dependencies = [
    "accelerate>=0.29.0",
    "bitsandbytes>=0.43.0",
    "huggingface-hub",
    "pandas",
    "peft>=0.10.0",
    "python-dotenv",
    "pyyaml",
    "scikit-learn",
    "torch>=2.1.0",
    "tqdm",
    "transformers>=4.40.0",
    "wandb"
]

[tool.setuptools.packages.find]
where = ["."]  # This tells python that "src" is where the code lives
include = ["src*"]